RVPS: rename proxy to grpc, server to native
